// client/Dao.types.ts
export interface Voter {
  addr: string;
  weight: number;
  [k: string]: unknown;
}